refactor e2e tests, add fork workflow tests, support fork workflow in GraphQL api
