+++ cypress/integration/editorial_workflow_spec_test_backend.js
@@ -1,0 +1,126 @@
+import '../utils/dismiss-local-backup';
+import {
+  login,
+  createPost,
+  createPostAndExit,
+  exitEditor,
+  goToWorkflow,
+  goToCollections,
+  updateWorkflowStatus,
+  publishWorkflowEntry,
+  assertWorkflowStatusInEditor,
+  assertPublishedEntry,
+  deleteEntryInEditor,
+  assertOnCollectionsPage,
+  assertEntryDeleted,
+  assertWorkflowStatus,
+  updateWorkflowStatusInEditor,
+  validateObjectFieldsAndExit,
+  validateNestedObjectFieldsAndExit,
+  validateListFieldsAndExit,
+} from '../utils/steps';
+import { setting1, setting2, workflowStatus, editorStatus } from '../utils/constants';
+
+const entry1 = {
+  title: 'first title',
+  body: 'first body',
+};
+const entry2 = {
+  title: 'second title',
+  body: 'second body',
+};
+const entry3 = {
+  title: 'third title',
+  body: 'third body',
+};
+
+describe('Test Backend Editorial Workflow', () => {
+  after(() => {
+    cy.task('restoreDefaults');
+  });
+
+  before(() => {
+    Cypress.config('defaultCommandTimeout', 4000);
+    cy.task('setupBackend', { backend: 'test' });
+  });
+
+  it('successfully loads', () => {
+    login();
+  });
+
+  it('can create an entry', () => {
+    login();
+    createPostAndExit(entry1);
+  });
+
+  it('can validate object fields', () => {
+    login();
+    validateObjectFieldsAndExit(setting1);
+  });
+
+  it('can validate fields nested in an object field', () => {
+    login();
+    validateNestedObjectFieldsAndExit(setting1);
+  });
+
+  it('can validate list fields', () => {
+    login();
+    validateListFieldsAndExit(setting2);
+  });
+
+  it('can publish an editorial workflow entry', () => {
+    login();
+    createPostAndExit(entry1);
+    goToWorkflow();
+    updateWorkflowStatus(entry1, workflowStatus.draft, workflowStatus.ready);
+    publishWorkflowEntry(entry1);
+  });
+
+  it('can change workflow status', () => {
+    login();
+    createPostAndExit(entry1);
+    goToWorkflow();
+    updateWorkflowStatus(entry1, workflowStatus.draft, workflowStatus.review);
+    updateWorkflowStatus(entry1, workflowStatus.review, workflowStatus.ready);
+    updateWorkflowStatus(entry1, workflowStatus.ready, workflowStatus.review);
+    updateWorkflowStatus(entry1, workflowStatus.review, workflowStatus.draft);
+    updateWorkflowStatus(entry1, workflowStatus.draft, workflowStatus.ready);
+  });
+
+  it('can change status on and publish multiple entries', () => {
+    login();
+    createPostAndExit(entry1);
+    createPostAndExit(entry2);
+    createPostAndExit(entry3);
+    goToWorkflow();
+    updateWorkflowStatus(entry3, workflowStatus.draft, workflowStatus.ready);
+    updateWorkflowStatus(entry2, workflowStatus.draft, workflowStatus.ready);
+    updateWorkflowStatus(entry1, workflowStatus.draft, workflowStatus.ready);
+    publishWorkflowEntry(entry3);
+    publishWorkflowEntry(entry2);
+    publishWorkflowEntry(entry1);
+    goToCollections();
+    assertPublishedEntry([entry3, entry2, entry1]);
+  });
+
+  it('can delete an entry', () => {
+    login();
+    createPost(entry1);
+    deleteEntryInEditor();
+    assertOnCollectionsPage();
+    assertEntryDeleted(entry1);
+  });
+
+  it('can update workflow status from within the editor', () => {
+    login();
+    createPost(entry1);
+    assertWorkflowStatusInEditor(editorStatus.draft);
+    updateWorkflowStatusInEditor(editorStatus.review);
+    assertWorkflowStatusInEditor(editorStatus.review);
+    updateWorkflowStatusInEditor(editorStatus.ready);
+    assertWorkflowStatusInEditor(editorStatus.ready);
+    exitEditor();
+    goToWorkflow();
+    assertWorkflowStatus(entry1, workflowStatus.ready);
+  });
+});
