+++ cypress/plugins/github.js
@@ -5,9 +5,7 @@ const simpleGit = require('simple-git/promise');
 
 const GIT_SSH_COMMAND = 'ssh -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no';
 
-function getGitHubClient() {
-  const { token } = getEnvs();
-
+function getGitHubClient(token) {
   const client = new Octokit({
     auth: `token ${token}`,
     baseUrl: 'https://api.github.com',
@@ -20,13 +18,15 @@ function getEnvs() {
     GITHUB_REPO_OWNER: owner,
     GITHUB_REPO_NAME: repo,
     GITHUB_REPO_TOKEN: token,
+    GITHUB_FORK_WORKFLOW_OWNER: forkOwner,
+    GITHUB_FORK_WORKFLOW_TOKEN: forkToken,
   } = process.env;
-  if (!owner || !repo || !token) {
+  if (!owner || !repo || !token || !forkOwner || !forkToken) {
     throw new Error(
-      'Please set GITHUB_REPO_OWNER, GITHUB_REPO_NAME, GITHUB_REPO_TOKEN environment variables',
+      'Please set GITHUB_REPO_OWNER, GITHUB_REPO_NAME, GITHUB_REPO_TOKEN, GITHUB_FORK_WORKFLOW_OWNER, GITHUB_FORK_WORKFLOW_TOKEN  environment variables',
     );
   }
-  return { owner, repo, token };
+  return { owner, repo, token, forkOwner, forkToken };
 }
 
 async function prepareTestGitHubRepo() {
@@ -34,7 +34,7 @@ async function prepareTestGitHubRepo() {
 
   const testRepoName = `${repo}-${Date.now()}`;
 
-  const client = getGitHubClient();
+  const client = getGitHubClient(token);
 
   console.log('Creating repository', testRepoName);
   await client.repos.createForAuthenticatedUser({
@@ -63,25 +63,47 @@ async function prepareTestGitHubRepo() {
   return { owner, repo: testRepoName, tempDir };
 }
 
-async function getUser() {
-  const { token } = getEnvs();
-  const client = getGitHubClient();
+async function getAuthenticatedUser(token) {
+  const client = getGitHubClient(token);
   const { data: user } = await client.users.getAuthenticated();
   return { ...user, token, backendName: 'github' };
 }
 
-async function deleteRepository({ owner, repo, tempDir }) {
-  console.log('Deleting repository', repo);
-  const client = getGitHubClient();
+async function getUser() {
+  const { token } = getEnvs();
+  return getAuthenticatedUser(token);
+}
+
+async function getForkUser() {
+  const { forkToken } = getEnvs();
+  return getAuthenticatedUser(forkToken);
+}
+
+async function deleteRepositories({ owner, repo, tempDir }) {
+  const { forkOwner, token, forkToken } = getEnvs();
+
+  console.log('Deleting repository', `${owner}/${repo}`);
+
+  let client = getGitHubClient(token);
   await client.repos.delete({
     owner,
     repo,
   });
+
   await fs.remove(tempDir);
+
+  client = getGitHubClient(forkToken);
+  console.log('Deleting forked repository', `${forkOwner}/${repo}`);
+  await client.repos.delete({
+    owner: forkOwner,
+    repo,
+  });
 }
 
-async function resetRepository({ owner, repo, tempDir }) {
-  const client = getGitHubClient();
+async function resetOriginRepo({ owner, repo, tempDir }) {
+  console.log('Resetting origin repo:', `${owner}/repo`);
+  const { token } = getEnvs();
+  const client = getGitHubClient(token);
 
   const { data: prs } = await client.pulls.list({
     repo,
@@ -100,7 +122,8 @@ async function resetRepository({ owner, repo, tempDir }) {
     ),
   );
 
-  const refs = prs.map(pr => `heads/${pr.head.ref}`);
+  const { data: branches } = await client.repos.listBranches({ owner, repo });
+  const refs = branches.filter(b => b.name !== 'master').map(b => `heads/${b.name}`);
 
   console.log('Deleting refs', refs);
   await Promise.all(
@@ -116,6 +139,42 @@ async function resetRepository({ owner, repo, tempDir }) {
   console.log('Resetting master');
   const git = simpleGit(tempDir).env({ ...process.env, GIT_SSH_COMMAND });
   await git.push(['--force', 'origin', 'master']);
+  console.log('Done resetting origin repo:', `${owner}/repo`);
 }
 
-module.exports = { prepareTestGitHubRepo, deleteRepository, getUser, resetRepository };
+async function resetForkedRepo({ repo }) {
+  const { forkToken, forkOwner } = getEnvs();
+  const client = getGitHubClient(forkToken);
+
+  const { data: repos } = await client.repos.list();
+  if (repos.some(r => r.name === repo)) {
+    console.log('Resetting forked repo:', `${forkOwner}/${repo}`);
+    const { data: branches } = await client.repos.listBranches({ owner: forkOwner, repo });
+    const refs = branches.filter(b => b.name !== 'master').map(b => `heads/${b.name}`);
+
+    console.log('Deleting refs', refs);
+    await Promise.all(
+      refs.map(ref =>
+        client.git.deleteRef({
+          owner: forkOwner,
+          repo,
+          ref,
+        }),
+      ),
+    );
+    console.log('Done resetting forked repo:', `${forkOwner}/repo`);
+  }
+}
+
+async function resetRepositories({ owner, repo, tempDir }) {
+  await resetOriginRepo({ owner, repo, tempDir });
+  await resetForkedRepo({ repo });
+}
+
+module.exports = {
+  prepareTestGitHubRepo,
+  deleteRepositories,
+  getUser,
+  getForkUser,
+  resetRepositories,
+};
